refactor optimization_utils into optimization_transfers and optimization_squad
