+++ airsenal/framework/utils.py
@@ -1256,28 +1256,3 @@ T = TypeVar("T")
 def fastcopy(obj: T) -> T:
     """faster replacement for copy.deepcopy()"""
     return loads(dumps(obj, -1))
-
-
-def get_discount_factor(next_gw, pred_gw, discount_type="exp", discount=14 / 15):
-    """
-    given the next gw and a predicted gw, retrieve discount factor. Either:
-        - exp: discount**n_ahead (discount reduces each gameweek)
-        - const: 1-(1-discount)*n_ahead (constant discount each gameweek, goes to
-          zero at gw 15 with default discount)
-    """
-    allowed_types = ["exp", "const", "constant"]
-    if discount_type not in allowed_types:
-        raise Exception("unrecognised discount type, should be exp or const")
-
-    if not next_gw:
-        # during tests 'none' is passed as the root gw, default to zero so the
-        # optimisation is done solely on pred_gw ahead.
-        next_gw = pred_gw
-    n_ahead = pred_gw - next_gw
-
-    if discount_type in ["exp"]:
-        score = discount ** n_ahead
-    elif discount_type in ["const", "constant"]:
-        score = max(1 - (1 - discount) * n_ahead, 0)
-
-    return score
